pyproject.toml
[tool.poetry]
name = "andrew"
version = "0.1.0"
description = ""
authors = ["Andrew <ajohnson0764@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pandas = "~=2.1"
hydra-core = "~=1.3"
pydantic = "~=2.5"
fsspec = {version = "~=2023.12", extras = ["gcs"]}
torch = "~=2.1.2"
lightning = "~=2.1.2"
torchmetrics = "~=1.3"
transformers = {version = "~=4.36", extras = ["torch"]}
mlflow = "~=2.8"
psycopg2-binary = "~=2.9"
google-cloud-compute = "~=1.15"
matplotlib = "~=3.7"
google-cloud-secret-manager = "~=2.16"
python-etcd = "~=0.4"
fastapi = "~=0.104"
uvicorn = {version = "~=0.24", extras=["standard"]}

[tool.poetry.group.dev.dependencies]
jupyterlab = "~=4.0"
pytest = "~=7.4"
black = "~=23.7.0"
isort = "~=5.12"
flake8 = "~=6.0"
mypy = "~=1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
indent = 4
lines_between_types = 1
known_first_party = ["cybulde"]
default_section = "THIRDPARTY"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.10"
show_error_codes = true
ignore_missing_imports = true
install_types = true
non_interactive = true
follow_imports = "silent"
no_strict_optional = true

warn_redundant_casts = false
warn_unused_ignores = true
warn_unused_configs = true
warn_return_any = true
warn_no_return = true
warn_incomplete_stub = true

disallow_subclassing_any = true

disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
check_untyped_defs = true
strict_optional = true

verbosity = 0
